requirements.txt
blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask==3.0.0
gitdb==4.0.10
GitPython==3.1.37
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
smmap==5.0.1
Werkzeug==3.0.0
